index.js cleared
